refactoring bike add/remove logic
